Copy diffusion from simulation-templates.
There is a small fix in `diffusion.h`. Kalium is correctly defined
with resolution 25.
