+++ demo/diffusion/README.md
@@ -1,0 +1,1 @@
+# Diffusion Example
