+++ demo/diffusion/src/diffusion.cc
@@ -12,6 +12,6 @@
 //
 // -----------------------------------------------------------------------------
 
-#include "diffusion_module.h"
+#include "diffusion.h"
 
 int main(int argc, const char** argv) { return bdm::Simulate(argc, argv); }
