+++ demo/diffusion/src/diffusion.h
@@ -1,0 +1,76 @@
+// -----------------------------------------------------------------------------
+//
+// Copyright (C) The BioDynaMo Project.
+// All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+//
+// See the LICENSE file distributed with this work for details.
+// See the NOTICE file distributed with this work for additional information
+// regarding copyright ownership.
+//
+// -----------------------------------------------------------------------------
+
+#ifndef DEMO_DIFFUSION_MODULE_H_
+#define DEMO_DIFFUSION_MODULE_H_
+
+#include <vector>
+
+#include "biodynamo.h"
+#include "diffusion_biology_modules.h"
+
+namespace bdm {
+
+// -----------------------------------------------------------------------------
+// This model creates 8 cells at each corner of a cube, and one in the middle.
+// The cell in the middle secretes a substance. The cells are modeled to
+// move according to the extracellular gradient; in this case to the middle.
+// -----------------------------------------------------------------------------
+
+// Define compile time parameter
+template <typename Backend>
+struct CompileTimeParam : public DefaultCompileTimeParam<Backend> {
+  using BiologyModules = Variant<Chemotaxis, KaliumSecretion>;
+};
+
+inline int Simulate(int argc, const char** argv) {
+  // Initialize BioDynaMo
+  Simulation<> simulation(argc, argv);
+
+  auto construct = [](const std::array<double, 3>& position) {
+    Cell cell(position);
+    cell.SetDiameter(30);
+    cell.SetMass(1.0);
+    cell.AddBiologyModule(Chemotaxis());
+    std::array<double, 3> secretion_position = {{50, 50, 50}};
+    if (position == secretion_position) {
+      cell.AddBiologyModule(KaliumSecretion());
+    }
+    return cell;
+  };
+  std::vector<std::array<double, 3>> positions;
+  positions.push_back({0, 0, 0});
+  positions.push_back({100, 0, 0});
+  positions.push_back({0, 100, 0});
+  positions.push_back({0, 0, 100});
+  positions.push_back({0, 100, 100});
+  positions.push_back({100, 0, 100});
+  positions.push_back({100, 100, 0});
+  positions.push_back({100, 100, 100});
+  // the cell responsible for secretion
+  positions.push_back({50, 50, 50});
+  ModelInitializer::CreateCells(positions, construct);
+
+  // Define the substances that cells may secrete
+  ModelInitializer::DefineSubstance(kKalium, "Kalium", 0.4, 0, 25);
+
+  // Run simulation for N timesteps
+  simulation.GetScheduler()->Simulate(350);
+  std::cout << "Simulation completed successfully!\n";
+  return 0;
+}
+
+}  // namespace bdm
+
+#endif  // DEMO_DIFFUSION_MODULE_H_
